backend/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
greenlet==3.3.2
grpcio==1.80.0
h11==0.16.0
h5py==3.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
keras==3.14.0
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.9.0.80
opt_einsum==3.4.0
optree==0.19.0
orjson==3.11.8
packaging==26.0
passlib==1.7.4
pillow==10.3.0
protobuf==4.25.9
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.33.1
rich==14.3.3
rich-toolkit==0.19.7
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.49
starlette==0.37.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.29.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.8
wrapt==2.1.2
